- Finishing the import of animations using assimp.

+++ Mythology/GameEngine/Geometry/Animation/SkinnedData.h
@@ -11,7 +11,7 @@ namespace GameEngine
 	{
 	public:
 		SkinnedData() = default;
-		SkinnedData(const std::vector<std::int32_t>& boneHierarchy, const std::vector<Eigen::Affine3f>& boneOffsets, const std::unordered_map<std::string, AnimationClip>& animations);
+		SkinnedData(const std::vector<std::int8_t>& boneHierarchy, const std::vector<Eigen::Affine3f>& boneOffsets, const std::unordered_map<std::string, AnimationClip>& animations);
 
 	public:
 		float GetClipStartTime(const std::string& clipName) const;
@@ -20,7 +20,7 @@ namespace GameEngine
 		void GetFinalTransforms(const std::string& clipName, float timePosition, std::vector<Eigen::Affine3f>& finalTransforms) const;
 
 	private:
-		std::vector<std::int32_t> m_boneHierarchy;
+		std::vector<std::int8_t> m_boneHierarchy;
 		std::vector<Eigen::Affine3f> m_boneOffsets;
 		std::unordered_map<std::string, AnimationClip> m_animations;
 	};
